+++ errors/retryable.go
@@ -42,6 +42,9 @@ func IsRetryableHTTP(err error) (retryType string, isRetryable bool) {
 // IsRetryableNetwork returns if the provided error is a retryable network related error. It also returns the
 // type, in string form, for optional logging and metrics use.
 func IsRetryableNetwork(err error) (retryType string, isRetryable bool) {
+	if IsRetryable(err) {
+		return "retryable", true
+	}
 	if IsTemporary(err) {
 		return "temporary", true
 	}
@@ -49,6 +52,18 @@ func IsRetryableNetwork(err error) (retryType string, isRetryable bool) {
 		return "timeout", true
 	}
 	return IsTemporaryConnection(err)
+}
+
+// IsRetryable returns true if the provided error is considered retryable by testing if it
+// complies with an interface implementing `Retryable() bool` or `IsRetryable bool` and calling the function.
+func IsRetryable(err error) bool {
+	var t interface{ IsRetryable() bool }
+	if errors.As(err, &t) && t.IsRetryable() {
+		return true
+	}
+
+	var t2 interface{ Retryable() bool }
+	return errors.As(err, &t2) && t2.Retryable()
 }
 
 // IsTemporary returns true if the provided error is considered retryable temporary error by testing if it