+++ net/http/helpers.go
@@ -4,6 +4,7 @@ import (
 	"compress/gzip"
 	"encoding/json"
 	"encoding/xml"
+	"errors"
 	"io"
 	"mime"
 	"net"
@@ -12,6 +13,7 @@ import (
 	"path/filepath"
 	"strings"
 
+	bytesext "github.com/go-playground/pkg/v5/bytes"
 	ioext "github.com/go-playground/pkg/v5/io"
 )
 
@@ -210,7 +212,7 @@ func DecodeMultipartForm(r *http.Request, qp QueryParamsOption, maxMemory int64,
 }
 
 // DecodeJSON decodes the request body into the provided struct and limits the request size via
-// an ioext.LimitReader using the maxMemory param.
+// an ioext.LimitReader using the maxBytes param.
 //
 // The Content-Type e.g. "application/json" and http method are not checked.
 //
@@ -244,7 +246,7 @@ func decodeJSON(headers http.Header, body io.Reader, qp QueryParamsOption, value
 }
 
 // DecodeXML decodes the request body into the provided struct and limits the request size via
-// an ioext.LimitReader using the maxMemory param.
+// an ioext.LimitReader using the maxBytes param.
 //
 // The Content-Type e.g. "application/xml" and http method are not checked.
 //
@@ -295,7 +297,7 @@ const (
 // Decode takes the request and attempts to discover its content type via
 // the http headers and then decode the request body into the provided struct.
 // Example if header was "application/json" would decode using
-// json.NewDecoder(ioext.LimitReader(r.Body, maxMemory)).Decode(v).
+// json.NewDecoder(ioext.LimitReader(r.Body, maxBytes)).Decode(v).
 //
 // This default to parsing query params if includeQueryParams=true and no other content type matches.
 //
@@ -319,6 +321,27 @@ func Decode(r *http.Request, qp QueryParamsOption, maxMemory int64, v interface{
 		if qp == QueryParams {
 			err = DecodeQueryParams(r, v)
 		}
+	}
+	return
+}
+
+// DecodeResponseAny takes the response and attempts to discover its content type via
+// the http headers and then decode the request body into the provided type.
+//
+// Example if header was "application/json" would decode using
+// json.NewDecoder(ioext.LimitReader(r.Body, maxBytes)).Decode(v).
+func DecodeResponseAny(r *http.Response, maxMemory bytesext.Bytes, v interface{}) (err error) {
+	typ := r.Header.Get(ContentType)
+	if idx := strings.Index(typ, ";"); idx != -1 {
+		typ = typ[:idx]
+	}
+	switch typ {
+	case nakedApplicationJSON:
+		err = decodeJSON(r.Header, r.Body, NoQueryParams, nil, maxMemory, v)
+	case nakedApplicationXML:
+		err = decodeXML(r.Header, r.Body, NoQueryParams, nil, maxMemory, v)
+	default:
+		err = errors.New("unsupported content type")
 	}
 	return
 }